+++ ltexls/src/main/java/org/bsplines/ltexls/settings/Settings.java
@@ -26,7 +26,7 @@ import org.eclipse.lsp4j.DiagnosticSeverity;
 
 public class Settings {
   private static final Set<String> defaultEnabled =
-      new HashSet<>(Arrays.asList("bibtex", "latex", "markdown", "rsweave"));
+      new HashSet<>(Arrays.asList("bibtex", "latex", "markdown", "restructuredtext", "rsweave"));
 
   private @Nullable Set<String> enabled;
   private @Nullable String languageShortCode;