+++ ltexls/src/test/java/org/bsplines/ltexls/parsing/restructuredtext/RestructuredtextFragmentizerTest.java
@@ -1,0 +1,57 @@
+/* Copyright (C) 2020 Julian Valentin, LTeX Development Community
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+package org.bsplines.ltexls.parsing.restructuredtext;
+
+import java.util.List;
+import org.bsplines.ltexls.parsing.CodeFragment;
+import org.bsplines.ltexls.parsing.CodeFragmentizer;
+import org.bsplines.ltexls.settings.Settings;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class RestructuredtextFragmentizerTest {
+  private static void testFragmentizer(CodeFragmentizer fragmentizer, String code) {
+    List<CodeFragment> codeFragments = fragmentizer.fragmentize(code, new Settings());
+    Assertions.assertEquals(3, codeFragments.size());
+
+    Assertions.assertEquals("restructuredtext", codeFragments.get(0).getCodeLanguageId());
+    Assertions.assertEquals(0, codeFragments.get(0).getFromPos());
+    Assertions.assertEquals(11, codeFragments.get(0).getCode().length());
+    Assertions.assertEquals("en-US", codeFragments.get(0).getSettings().getLanguageShortCode());
+
+    Assertions.assertEquals("restructuredtext", codeFragments.get(1).getCodeLanguageId());
+    Assertions.assertEquals(11, codeFragments.get(1).getFromPos());
+    Assertions.assertEquals(37, codeFragments.get(1).getCode().length());
+    Assertions.assertEquals("de-DE", codeFragments.get(1).getSettings().getLanguageShortCode());
+
+    Assertions.assertEquals("restructuredtext", codeFragments.get(2).getCodeLanguageId());
+    Assertions.assertEquals(48, codeFragments.get(2).getFromPos());
+    Assertions.assertEquals(37, codeFragments.get(2).getCode().length());
+    Assertions.assertEquals("en-US", codeFragments.get(2).getSettings().getLanguageShortCode());
+  }
+
+  @Test
+  public void test() {
+    CodeFragmentizer fragmentizer = CodeFragmentizer.create("restructuredtext");
+
+    testFragmentizer(fragmentizer,
+        "Sentence 1\n"
+        + "\n.. ltex: language=de-DE\n\nSentence 2\n"
+        + "\n..\tltex:\tlanguage=en-US\n\nSentence 3\n");
+  }
+
+  @Test
+  public void testWrongSettings() {
+    RestructuredtextFragmentizer restructuredtextFragmentizer =
+        new RestructuredtextFragmentizer("restructuredtext");
+    Assertions.assertDoesNotThrow(() -> restructuredtextFragmentizer.fragmentize(
+        "Sentence 1\n\n.. ltex: languagede-DE\n\nSentence 2\n", new Settings()));
+    Assertions.assertDoesNotThrow(() -> restructuredtextFragmentizer.fragmentize(
+        "Sentence 1\n\n.. ltex: unknownKey=abc\n\nSentence 2\n", new Settings()));
+  }
+}